requirements.txt
Django==4.2.16
Pillow==10.4.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
requests==2.31.0
dj-database-url==2.1.0
